+++ src/control_plane/engine/volume_mount.rs
@@ -42,11 +42,11 @@ impl VolumeMount {
         })
     }
 
-    pub(super) fn source(&self) -> &str {
+    pub(crate) fn source(&self) -> &str {
         &self.source
     }
 
-    pub(super) fn target(&self) -> &str {
+    pub(crate) fn target(&self) -> &str {
         &self.target
     }